+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2026-02-16 18:23:40
+
+### Changed Files
+- `src/features/settings/SettingsDashboard.tsx`
+
+### Statistics
+```
+ 1 file changed, 8 insertions(+), 3 deletions(-)
+```
+
 ## [Checkpoint] - 2026-02-16 18:00:53
 
 ### Changed Files